onesignal-socket-layer/requirements.txt
onesignal-python-api==2.0.2
python-socketio==5.11.2
python-socketio[client]